requirements.txt
asgiref==3.8.1
Django==5.0.7
django-cors-headers==4.4.0
django-jazzmin==3.0.0
djangorestframework==3.15.2
mysqlclient==2.2.4
sqlparse==0.5.1
tzdata==2024.1
